requirements.txt
mistralai==0.0.20
streamlit==1.28.1
python-dotenv==1.0.0
